backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pathway==0.28.0
litellm==1.77.2.post1
pydantic==2.9.2
python-dotenv==1.0.1
pdf2image==1.17.0
unstructured==0.18.32
docling==2.74.0
torch==2.10.0+cpu
sentence-transformers==3.4.1

